requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
aiofiles==23.2.1
jinja2==3.1.4
pydantic==2.8.2
pydantic-settings==2.4.0
aiogram==3.10.0
python-dotenv==1.0.1
sqlalchemy==2.0.32
databases==0.6.1
